# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_assistant_stream_assistant_stream_post": {
        "properties": {
          "conversation_id": {
            "default": "",
            "title": "Conversation Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "Body_assistant_stream_assistant_stream_post",
        "type": "object"
      },
      "Body_settings_save_settings_post": {
        "properties": {
          "api_key": {
            "default": "",
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "max_tokens": {
            "default": "",
            "title": "Max Tokens",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "temperature": {
            "default": "",
            "title": "Temperature",
            "type": "string"
          }
        },
        "required": [
          "base_url",
          "model"
        ],
        "title": "Body_settings_save_settings_post",
        "type": "object"
      },
      "ConversationMessageView": {
        "description": "会话里的一条消息。",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ConversationMessageView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SettingsForm": {
        "description": "配置页提交的表单。`api_key` 留空表示「不改」。",
        "properties": {
          "api_key": {
            "default": "",
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "max_tokens": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          }
        },
        "required": [
          "base_url",
          "model"
        ],
        "title": "SettingsForm",
        "type": "object"
      },
      "SettingsView": {
        "description": "回给界面的配置（密钥只给掩码）。",
        "properties": {
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "masked_key": {
            "title": "Masked Key",
            "type": "string"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          }
        },
        "required": [
          "base_url",
          "model",
          "masked_key",
          "configured"
        ],
        "title": "SettingsView",
        "type": "object"
      },
...
      "StreamRequest": {
        "description": "对话请求体（流式与一次性共用）。",
        "properties": {
          "conversation_id": {
            "default": "",
            "title": "Conversation Id",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "StreamRequest",
        "type": "object"
      },
      "TurnResponse": {
        "description": "一次性对话的结果。",
        "properties": {
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "iterations": {
            "title": "Iterations",
            "type": "integer"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "conversation_id",
          "reply",
          "iterations",
          "truncated"
        ],
        "title": "TurnResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Hunter1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/assistant/conversations/{conversation_id}": {
      "get": {
        "operationId": "conversation_messages_api_assistant_conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationMessageView"
                  },
                  "title": "Response Conversation Messages Api Assistant Conversations  Conversation Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "会话消息"
      }
    },
    "/api/assistant/stream": {
      "post": {
        "operationId": "stream_api_assistant_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "流式对话（SSE）"
      }
    },
    "/api/assistant/turn": {
      "post": {
        "operationId": "turn_api_assistant_turn_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "跑一轮对话（一次性返回）"
      }
    },
...
    "/api/settings": {
      "get": {
        "operationId": "read_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SettingsView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Read Settings Api Settings Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "读取 LLM 配置"
      },
      "put": {
        "operationId": "save_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "保存 LLM 配置"
      }
    },
...
    "/assistant/stream": {
      "post": {
        "description": "流式回合：以 `text/event-stream` 边生成边推。\n\n与 `/assistant` 的差别只在交付方式 —— 内容逐段到达，工具执行前后各有\n一个事件。**失败也走事件流**（一条 `error` 事件）：响应一旦开始就没法\n再重定向，这是流式端点最容易漏掉的一条。",
        "operationId": "assistant_stream_assistant_stream_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_assistant_stream_assistant_stream_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assistant Stream"
      }
    },
...
    "/settings": {
      "get": {
        "operationId": "settings_page_settings_get",
        "parameters": [
          {
            "in": "query",
            "name": "saved",
            "required": false,
            "schema": {
              "default": "",
              "title": "Saved",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Settings Page"
      },
      "post": {
        "operationId": "settings_save_settings_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_settings_save_settings_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Settings Save"
      }
    },
...
  }
}